+++ test/tests/StaticVector/StaticVectorIterator.cpp
@@ -13,7 +13,7 @@ namespace FcfTest {
         static_vector_type v = {1,2,3};
         size_t key = 0;
         for(static_vector_type::iterator it = v.begin(); it != v.end(); ++it){
-          FCF_TEST(*it == (key+1), *it, (key+1));
+          FCF_TEST(*it == (int)(key+1), *it, (key+1));
           ++key;
         }
       }
@@ -22,7 +22,7 @@ namespace FcfTest {
         static_vector_type v = {1,2,3};
         size_t key = 0;
         for(int itm : v){
-          FCF_TEST(itm == (key+1), itm, (key+1));
+          FCF_TEST(itm == (int)(key+1), itm, (key+1));
           ++key;
         }
       }
@@ -31,8 +31,8 @@ namespace FcfTest {
         static_vector_type v = {1,2,3};
         size_t key = 0;
         for(static_vector_type::const_iterator it = v.cbegin(); it != v.cend(); ++it){
-          FCF_TEST(*it == (key+1), *it, (key+1));
-          FCF_TEST(*(it.operator->()) == (key+1), *it, (key+1));
+          FCF_TEST(*it == (int)(key+1), *it, (key+1));
+          FCF_TEST(*(it.operator->()) == (int)(key+1), *it, (key+1));
           ++key;
         }
       }
